kafka/requirements.txt
kafka-python==2.0.2
minio==7.2.5
requests==2.31.0
python-dotenv==1.0.0